make config private
