+++ pkg/collector/corechecks/snmp/snmp_test.go
@@ -557,159 +557,6 @@ profiles:
 	sender.AssertServiceCheck(t, "snmp.can_check", metrics.ServiceCheckOK, "", snmpTags, "")
 }
 
-func TestProfileWithSysObjectIdDetection(t *testing.T) {
-	checkconfig.SetConfdPathAndCleanProfiles()
-	sess := session.CreateMockSession()
-	chk := Check{}
-	// language=yaml
-	rawInstanceConfig := []byte(`
-ip_address: 1.2.3.4
-community_string: public
-`)
-	// language=yaml
-	rawInitConfig := []byte(`
-profiles:
-  f5-big-ip:
-    definition_file: f5-big-ip.yaml
-`)
-
-	err := chk.Configure(rawInstanceConfig, rawInitConfig, "test")
-	assert.Nil(t, err)
-	chk.singleDeviceCk.SetSession(sess)
-
-	sender := mocksender.NewMockSender(chk.ID()) // required to initiate aggregator
-	sender.On("Gauge", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return()
-	sender.On("MonotonicCount", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return()
-	sender.On("ServiceCheck", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return()
-	sender.On("Commit").Return()
-
-	sysObjectIDPacket := gosnmp.SnmpPacket{
-		Variables: []gosnmp.SnmpPDU{
-			{
-				Name:  "1.3.6.1.2.1.1.2.0",
-				Type:  gosnmp.ObjectIdentifier,
-				Value: "1.3.6.1.4.1.3375.2.1.3.4.1",
-			},
-		},
-	}
-
-	packet := gosnmp.SnmpPacket{
-		Variables: []gosnmp.SnmpPDU{
-			{
-				Name:  "1.3.6.1.2.1.1.3.0",
-				Type:  gosnmp.TimeTicks,
-				Value: 20,
-			},
-			{
-				Name:  "1.3.6.1.2.1.1.5.0",
-				Type:  gosnmp.OctetString,
-				Value: []byte("foo_sys_name"),
-			},
-			{
-				Name:  "1.3.6.1.4.1.3375.2.1.1.2.1.44.0",
-				Type:  gosnmp.Integer,
-				Value: 30,
-			},
-		},
-	}
-
-	bulkPacket := gosnmp.SnmpPacket{
-		Variables: []gosnmp.SnmpPDU{
-			{
-				Name:  "1.3.6.1.2.1.2.2.1.13.1",
-				Type:  gosnmp.Integer,
-				Value: 131,
-			},
-			{
-				Name:  "1.3.6.1.2.1.2.2.1.14.1",
-				Type:  gosnmp.Integer,
-				Value: 141,
-			},
-			{
-				Name:  "1.3.6.1.2.1.31.1.1.1.1.1",
-				Type:  gosnmp.OctetString,
-				Value: []byte("nameRow1"),
-			},
-			{
-				Name:  "1.3.6.1.2.1.31.1.1.1.18.1",
-				Type:  gosnmp.OctetString,
-				Value: []byte("descRow1"),
-			},
-			{
-				Name:  "1.3.6.1.2.1.2.2.1.13.2",
-				Type:  gosnmp.Integer,
-				Value: 132,
-			},
-			{
-				Name:  "1.3.6.1.2.1.2.2.1.14.2",
-				Type:  gosnmp.Integer,
-				Value: 142,
-			},
-			{
-				Name:  "1.3.6.1.2.1.31.1.1.1.1.2",
-				Type:  gosnmp.OctetString,
-				Value: []byte("nameRow2"),
-			},
-			{
-				Name:  "1.3.6.1.2.1.31.1.1.1.18.2",
-				Type:  gosnmp.OctetString,
-				Value: []byte("descRow2"),
-			},
-			{
-				Name:  "9", // exit table
-				Type:  gosnmp.Integer,
-				Value: 999,
-			},
-			{
-				Name:  "9", // exit table
-				Type:  gosnmp.Integer,
-				Value: 999,
-			},
-			{
-				Name:  "9", // exit table
-				Type:  gosnmp.Integer,
-				Value: 999,
-			},
-			{
-				Name:  "9", // exit table
-				Type:  gosnmp.Integer,
-				Value: 999,
-			},
-		},
-	}
-
-	sess.On("Get", []string{"1.3.6.1.2.1.1.2.0"}).Return(&sysObjectIDPacket, nil)
-	sess.On("Get", []string{"1.3.6.1.2.1.1.3.0", "1.3.6.1.4.1.3375.2.1.1.2.1.44.0", "1.3.6.1.4.1.3375.2.1.1.2.1.44.999", "1.2.3.4.5", "1.3.6.1.2.1.1.5.0"}).Return(&packet, nil)
-	sess.On("GetBulk", []string{"1.3.6.1.2.1.2.2.1.13", "1.3.6.1.2.1.2.2.1.14", "1.3.6.1.2.1.31.1.1.1.1", "1.3.6.1.2.1.31.1.1.1.18"}, checkconfig.DefaultBulkMaxRepetitions).Return(&bulkPacket, nil)
-
-	err = chk.Run()
-	assert.Nil(t, err)
-
-	snmpTags := []string{"snmp_device:1.2.3.4", "snmp_profile:f5-big-ip", "device_vendor:f5", "snmp_host:foo_sys_name",
-		"some_tag:some_tag_value", "prefix:f", "suffix:oo_sys_name"}
-	row1Tags := append(common.CopyStrings(snmpTags), "interface:nameRow1", "interface_alias:descRow1")
-	row2Tags := append(common.CopyStrings(snmpTags), "interface:nameRow2", "interface_alias:descRow2")
-
-	sender.AssertMetric(t, "Gauge", "snmp.devices_monitored", float64(1), "", snmpTags)
-	sender.AssertMetric(t, "Gauge", "snmp.sysUpTimeInstance", float64(20), "", snmpTags)
-	sender.AssertMetric(t, "MonotonicCount", "snmp.ifInErrors", float64(141), "", row1Tags)
-	sender.AssertMetric(t, "MonotonicCount", "snmp.ifInErrors", float64(142), "", row2Tags)
-	sender.AssertMetric(t, "MonotonicCount", "snmp.ifInDiscards", float64(131), "", row1Tags)
-	sender.AssertMetric(t, "MonotonicCount", "snmp.ifInDiscards", float64(132), "", row2Tags)
-	sender.AssertMetric(t, "Gauge", "snmp.sysStatMemoryTotal", float64(30), "", snmpTags)
-
-	assert.Equal(t, false, chk.singleDeviceCk.Config.AutodetectProfile)
-
-	// Make sure we don't auto detect and add metrics twice if we already did that previously
-	firstRunMetrics := chk.singleDeviceCk.Config.Metrics
-	firstRunMetricsTags := chk.singleDeviceCk.Config.MetricTags
-	err = chk.Run()
-	assert.Nil(t, err)
-
-	assert.Len(t, chk.singleDeviceCk.Config.Metrics, len(firstRunMetrics))
-	assert.Len(t, chk.singleDeviceCk.Config.MetricTags, len(firstRunMetricsTags))
-}
-
 func TestServiceCheckFailures(t *testing.T) {
 	checkconfig.SetConfdPathAndCleanProfiles()
 	sess := session.CreateMockSession()
